requirements.txt
fastapi
uvicorn
transformers
torch
Pillow
python-multipart
accelerate
